refactor Entity files to directories
