+++ src/data/input_output/player_updaters.py
@@ -2,69 +2,13 @@ from abc import ABC, abstractmethod
 from collections.abc import Callable
 from dataclasses import dataclass
 from functools import cached_property
-from typing import Any, override
+from typing import override
 
 from common.i18n import _
 from data.player import Player
 from utils.entity import IdentifiableEntity
-from utils.enum import TrfType, TournamentRating
+from utils.enum import TournamentRating
 from database.sqlite.fide.fide_database import FideDatabase
-
-
-# ---------------------------------------------------------------------------------
-# Tournament exporters
-# ---------------------------------------------------------------------------------
-
-
-class AbstractTournamentExporter(ABC):
-    @property
-    @abstractmethod
-    def name(self) -> str:
-        """Represents the exporter in the UI"""
-        pass
-
-    @property
-    @abstractmethod
-    def download_route(self) -> str:
-        """Route downloading the export file.
-        Should take as parameters event_uniq_id: str and tournament_id: int"""
-
-    @property
-    def route_parameters(self) -> dict[str, Any]:
-        return {}
-
-
-class Trf16TournamentExporter(AbstractTournamentExporter):
-    @property
-    def download_route(self) -> str:
-        return 'admin-tournament-trf-export'
-
-    @property
-    def name(self) -> str:
-        return _('Export to TRF16 (rating)')
-
-    @property
-    def route_parameters(self) -> dict[str, Any]:
-        return {'usage': TrfType.RATING}
-
-
-class TrfBxTournamentExporter(AbstractTournamentExporter):
-    @property
-    def download_route(self) -> str:
-        return 'admin-tournament-trf-export'
-
-    @property
-    def name(self) -> str:
-        return _('Export to TRF(bx) (pairing)')
-
-    @property
-    def route_parameters(self) -> dict[str, Any]:
-        return {'usage': TrfType.PAIRING}
-
-
-# ---------------------------------------------------------------------------------
-# Player Updater
-# ---------------------------------------------------------------------------------
 
 
 @dataclass
