+++ src/data/print_documents/managers.py
@@ -1,0 +1,42 @@
+from data.print_documents import documents, options, player_splitters
+from data.print_documents.documents import PrintDocument
+from data.print_documents.options import PrintOption
+from data.print_documents.player_splitters import PlayerSplitter
+from plugins.manager import plugin_manager
+from utils.entity import EntityManager
+
+
+class PrintDocumentManager(EntityManager[PrintDocument]):
+    @staticmethod
+    def entity_types() -> list[type[PrintDocument]]:
+        return [
+            documents.PlayerListPrintDocument,
+            documents.PairingPrintDocument,
+            documents.ResultPrintDocument,
+            documents.PlayerRankingPrintDocument,
+            documents.PlayerCrosstablePrintDocument,
+        ]
+
+
+class PrintDocumentOptionManager(EntityManager[PrintOption]):
+    @staticmethod
+    def entity_types() -> list[type[options.PrintOption]]:
+        return [
+            options.RoundPrintOption,
+            options.PlayerSplitPrintOption,
+        ]
+
+
+class PrintPlayerSplitterManager(EntityManager[PlayerSplitter]):
+    @staticmethod
+    def entity_types() -> list[type[PlayerSplitter]]:
+        splitters = [
+            player_splitters.NoSplitPlayerSplitter,
+            player_splitters.CategoryPlayerSplitter,
+            player_splitters.ClubPlayerSplitter,
+            player_splitters.FederationPlayerSplitter,
+        ]
+        plugin_manager.hook.insert_print_player_splitter_types(
+            player_splitter_types=splitters
+        )
+        return splitters
