+++ src/data/tie_breaks/tie_breaks.py
@@ -5,16 +5,25 @@ from collections.abc import Iterable
 from contextlib import suppress
 from decimal import Decimal
 from math import isclose
-from types import UnionType
-from typing import Any, TYPE_CHECKING, Protocol, TypeVar, override
+from typing import TYPE_CHECKING, Protocol, TypeVar
 
 
 from common.i18n import _
 from data.pairing import Pairing
 from data.player import TournamentPlayer
+from data.tie_breaks.options import (
+    TieBreakOption,
+    CutTieBreakOption,
+    CutTopTieBreakOption,
+    CutBottomTieBreakOption,
+    PlayedModifierTieBreakOption,
+    ForeModifierTieBreakOption,
+    LimitTieBreakOption,
+    ExcludeIdsTieBreakOption,
+)
 from utils import StaticUtils
-from utils.entity import Option, OptionHandler, OptionError
 from utils.enum import BoardColor, Result, TournamentPairing
+from utils.option import OptionHandler, OptionError
 
 if TYPE_CHECKING:
     from data.player import Player
@@ -100,115 +109,7 @@ class TieBreakUtils:
         return dummy
 
 
-# ---------------------------------------------------------------------------------
-# Tie-break options
-# ---------------------------------------------------------------------------------
-
-class TieBreakOption(Option, ABC):
-    """Abstract class representing an option of a tie-break"""
-
-    @property
-    def template_name(self) -> str:
-        # TODO Implement templates for tie-break options
-        return ''
-
-
-class AbstractCutTieBreakOption(TieBreakOption, ABC):
-    @property
-    def type(self) -> type | UnionType:
-        return int
-
-    @property
-    def default_value(self) -> Any:
-        return 0
-
-    @override
-    def validate(self):
-        super().validate()
-        if self.value < 0:
-            raise OptionError(_('A positive integer is expected.'), self)
-
-
-class CutTieBreakOption(AbstractCutTieBreakOption):
-    @staticmethod
-    def static_id() -> str:
-        return 'CUT'
-
-
-class CutTopTieBreakOption(AbstractCutTieBreakOption):
-    @staticmethod
-    def static_id() -> str:
-        return 'CUT_TOP'
-
-
-class CutBottomTieBreakOption(AbstractCutTieBreakOption):
-    @staticmethod
-    def static_id() -> str:
-        return 'CUT_BOTTOM'
-
-
-class PlayedModifierTieBreakOption(TieBreakOption):
-    @staticmethod
-    def static_id() -> str:
-        return 'PLAYED_MODIFIER'
-
-    @property
-    def type(self) -> type | UnionType:
-        return bool
-
-    @property
-    def default_value(self) -> Any:
-        return False
-
-
-class ForeModifierTieBreakOption(TieBreakOption):
-    @staticmethod
-    def static_id() -> str:
-        return 'FORE_MODIFIER'
-
-    @property
-    def type(self) -> type | UnionType:
-        return bool
-
-    @property
-    def default_value(self) -> Any:
-        return False
-
-
-class LimitTieBreakOption(TieBreakOption):
-    @staticmethod
-    def static_id() -> str:
-        return 'LIMIT'
-
-    @property
-    def type(self) -> type | UnionType:
-        return float | None
-
-    @property
-    def default_value(self) -> Any:
-        return None
-
-
-class ExcludeIdsTieBreakOption(TieBreakOption):
-    @staticmethod
-    def static_id() -> str:
-        return 'EXCLUDE_IDS'
-
-    @property
-    def type(self) -> type | UnionType:
-        return Iterable[int] | None
-
-    @property
-    def default_value(self) -> Any:
-        return None
-
-
-# ---------------------------------------------------------------------------------
-# Tie-breaks
-# ---------------------------------------------------------------------------------
-
-
-class TieBreak(OptionHandler, ABC):
+class TieBreak(OptionHandler[TieBreakOption], ABC):
     """Abstract class representing a tie-break"""
 
     @property
@@ -434,7 +335,7 @@ class ProgressiveScoresTieBreak(TieBreak):
         return _('Progressive')
 
     @staticmethod
-    def available_options() -> list[type[Option]]:
+    def available_options() -> list[type[TieBreakOption]]:
         return [CutTieBreakOption]
 
     def compute_player_value(
